Split up gateway into multiple files
This also renames the SocketThreadErro
