+++ src/errors.rs
@@ -17,7 +17,7 @@ pub enum DiscordError {
     HttpError(surf::Exception),
     IoError(std::io::Error),
     HeartbeatSeqUpdateError(tokio::sync::watch::error::SendError<Option<u64>>),
-    SocketThread(tokio::sync::mpsc::error::SendError<crate::api::gateway::Payload>),
+    SocketThreadError(tokio::sync::mpsc::error::SendError<crate::api::gateway::Payload>),
     HeartbeatTimeError(tokio::sync::watch::error::SendError<std::time::Instant>),
     ApiError(crate::api::ApiError),
     GatewayError(GatewayError),
@@ -35,7 +35,9 @@ impl std::fmt::Display for DiscordError {
                 write!(f, "Heartbeat sequence update error: {}", e)
             }
             Self::HeartbeatTimeError(ref e) => write!(f, "Heartbeat time update error: {}", e),
-            Self::SocketThread(ref e) => write!(f, "Socket thread communcation failiure: {}", e),
+            Self::SocketThreadError(ref e) => {
+                write!(f, "Socket thread communcation failiure: {}", e)
+            }
             Self::GatewayError(ref e) => write!(f, "Gateway error: {}", e),
         }
     }
@@ -53,7 +55,7 @@ convert_error!(
 convert_error!(
     tokio::sync::mpsc::error::SendError<crate::api::gateway::Payload>,
     DiscordError,
-    SocketThread
+    SocketThreadError
 );
 convert_error!(GatewayError, DiscordError, GatewayError);
